Unify WMMA and FPU operator typevars [NFC]
The WMMA operator only had a T typevar for the accumulator type,
while the FPU operator had DT for the destination type and CT for
the compute type. Unify that by adding both compute type (CT) and
accumulator type (AT) typevars that indicate the type that should
be used for the register-level storage and operations.

Note that the WMMA operator's typevars are actually not useful, and
should match the eltype of the shared memory (as we use WMMA intrinsics
to load/store shared memory, so cannot convert between shared memory and
registers). However, we need the accumulator typevar as it cannot be
inferred from arguments at some points, so I decided to add the compute
typevar too for alignment with the FPU operator.