Use a zero layout for C in shared memory if beta=0
While we already avoid a global load from C in case beta == 0, we still
emit stores to shared memory and loads from shared memory for C.

Instead, we should also use a zero layout for C in shared memory, which
eliminates these extra loads and stores.

This does not seem to influence the performance of GEMM, even for small
matrices, or highly rectangular GEMMs with small K, but it does make a
difference for some TCs, I've noticed, so let's do this, anyway.