Filter out zero ratios when calculating geomean
E.g. in case we do not find a configuration for a certain TC, the ratio
is cuTENOR_time / Inf = 0, leading to a 0 geomean. While we should fix
such cases, let's calculate the geomean w/o taking these into account to
get an idea of the average performance for now.